requirements.txt
alembic==1.1.0
blinker==1.4
certifi==2019.6.16
chardet==3.0.4
Click==7.0
dnspython==1.16.0
eventlet==0.25.1
Flask==1.1.1
Flask-Dance==2.2.0
Flask-DebugToolbar==0.10.1
Flask-DotEnv==0.1.2
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-SeaSurf==0.2.2
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.4.0
flask-talisman==0.7.0
greenlet==0.4.15
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
MarkupSafe==1.1.1
monotonic==1.5
oauthlib==3.1.0
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
python-engineio==3.9.3
python-socketio==4.3.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
SQLAlchemy==1.3.8
urllib3==1.25.3
URLObject==2.4.3
Werkzeug==0.15.5
